+++ src/main/java/com/huy/crm/dao/CustomerDAO.java
@@ -9,8 +9,8 @@ import java.util.List;
 public interface CustomerDAO {
     List<Customer> getCustomers(CustomerParams customerParams);
     int getCustomersCount(CustomerParams customerParams);
-    Customer getCustomerById(int id);
+    Customer getCustomerById(long id);
     Customer getCustomerByEmail(String email);
     void saveCustomer(Customer customer);
-    void deleteCustomer(int id);
+    void deleteCustomer(long id);
 }
